Improve several error codes so attribute access of names is recognised (#161)
Several error codes had issues due to the fact that the plugin hardcoded names such as `Literal` or `Union` into the logic, but did not consider that a user might access these names via attribute access, e.g. `typing_extensions.Literal` or `typing.Union`. For example, the plugin recognised `X: TypeAlias = int` as a valid type alias, but did not recognise `X: typing.TypeAlias = int` as a valid type alias. It's not standard "typeshed style" to access names in that way, but it's nonetheless valid, so we should probably account for that possibility.

This commit fixes these bugs.