Support `typing_extensions.overload` and `typing_extensions.NamedTuple` (#231)
Y028 currently isn't emitted if somebody uses assignment-based syntax for `typing_extensions.NamedTuple` in a stub. `typing_extensions.NamedTuple` doesn't exist yet, but it will exist in the next release of `typing_extensions`, and in the future we might want to use it in stubs for generic `NamedTuple`s.

Y034 has some exemptions for methods decorated with `@typing.overload`. Make sure those exemptions also work for `typing_extensions.overload`.

Make it legal to annotate the first argument in an `__(a)exit__` method with `builtins.BaseException` (as opposed to just `BaseException`).