Introduce Y052: Disallow assignments to constant values in global or class namespaces where the assignments don't have type annotations (#339)
Following #326, we now allow constructs such as these in a stub file:

```python
x = 1
y = "foo"
z = b"bar"
```

It's good that we now allow variables to have default values. However,
the above constructs are problematic in a stub, as we're leaving it up
to the type checker to make inferences about the types of these
variables, which can have unpredictable consequences. For example, the
`x` variable could be inferred to be of type `int`, `Final[int]`,
`Literal[1]`, or `Final[Literal[1]]`. In a class namespace, the problem
is even worse -- the `eggs` variable in the following snippet could
reasonably be inferred as being of type `str`, `ClassVar[str]`,
`Final[str]`, `Literal["EGGS"]`, `ClassVar[Literal["EGGS"]]`, or
`Final[Literal["EGGS"]]`:

```python
class Whatever:
    eggs = "EGGS"
```

This PR introduces Y052, enforcing type annotations for assignments to
simple constants.