Simplify Y033 check (#376)
Rather than opening the file and splitting it into a list of lines
ourselves, we can just ask flake8 directly to provide us with the list
of lines in the file we're checking (similar to how we already do with
the AST tree). It's one of the data points that flake8 details we're
allowed to directly request in [their docs on writing
plugins](https://flake8.pycqa.org/en/latest/plugin-development/plugin-parameters.html#indicating-desired-data).