Move to `hatch-vcs` for versioning (#422)
This means there's one less thing to do manually when cutting a new
release. Rather than us manually updating the `__version__` attribute at
the top of the `pyi.py` file, `hatch-vcs` looks at the git tag to
determine the version that should be added to the metadata when building
a wheel.