Fix typeshed_primer workflow
It's been broken since we switched the workflow to use `uv`, because `uv` doesn't (by default) reinstall an editable package if it sees that it's already installed with the "right version". (I personally find this pretty surprising behaviour for editable installs. It's already tracked in https://github.com/astral-sh/uv/issues/2844.)